requirements.txt
streamlit
google-generativeai
requests
pandas
mysql-connector-python